Add files via upload
This repository displays the correlation of fatalities in the workspace for each state at the State and Federal level based on a 2012 data pull. The data was provided by the Census and added as a csv file. My objective was massage the data and upload it into pbix format to display the correlations