Add files via upload
The datasets that were used were based on professional baseball players and their statistics from the past and future. When using this data, I focused on the common statistical correlations of baseball such as home runs to strike out and stolen bases to caught stealing ratios. Sports is a great theme to analyze to create statistical data, while baseball can provide a lot of data that can be utilized, the overall data can be skewed based on the strategic scenarios of the game.